# Conflicts:
#	crates/backend-uzu/src/encodable_block/classifier.rs
#	crates/backend-uzu/src/encodable_block/decoder.rs
#	crates/backend-uzu/src/encodable_block/transformer.rs
#	crates/backend-uzu/src/encodable_block/transformer_layer.rs
#	crates/backend-uzu/src/engine/classifier_model.rs

+++ crates/backend-uzu/src/backends/cpu/kernel/embedding/quant_embedding.rs
@@ -33,7 +33,7 @@ pub fn quantized_embedding_lookup<T: ArrayElement + Float>(
         unimplemented!("not supported yet");
     }
 
-    let packing_divisor = quantization_mode.packing_divisor() as u32;
+    let packing_divisor = quantization_mode.packing_divisor();
     let weights_stride = model_dim / packing_divisor;
     let num_groups = model_dim.div_ceil(group_size);
     let zero_points_stride = match quantization_mode {
