+++ crates/backend-uzu/src/engine/language_model/mod.rs
@@ -119,7 +119,7 @@ impl<B: Backend> Engine<B> {
         let generation_config = config.generation_config;
 
         #[cfg(grammar)]
-        let vocab_size = config.decoder_config.vocab_size;
+        let vocab_size = config.decoder_config.vocab_size as usize;
 
         Ok(LanguageModel {
             engine: self.clone(),
@@ -136,11 +136,11 @@ impl<B: Backend> Engine<B> {
 }
 
 impl<B: Backend> LanguageModel<B> {
-    pub fn max_context_length(&self) -> Option<usize> {
+    pub fn max_context_length(&self) -> Option<u32> {
         self.decoder.max_context_length()
     }
 
-    pub fn recommended_context_length(&self) -> Option<usize> {
+    pub fn recommended_context_length(&self) -> Option<u32> {
         let max_context_length = self.max_context_length();
 
         // TODO: This is not the correct way to do it, there should be a real memory model
@@ -160,7 +160,7 @@ impl<B: Backend> LanguageModel<B> {
                 16384
             };
 
-            Some(usize::min(max_context_length, platform_recommended_context_length))
+            Some(u32::min(max_context_length, platform_recommended_context_length))
         } else {
             // We just assume that unlimited context means constant state size on all mixers and is thus free
             None
